+++ cli/backup/chain/points_test.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/require"
-	"github.com/tarantool/tt/lib/backup"
+	"github.com/tarantool/tt/cli/backup"
 )
 
 func TestClusterPointsJoinReplicasetsByName(t *testing.T) {