+++ test/integration/completion/test_completion.py
@@ -48,6 +48,7 @@ tt_root_command: set[str] = set(
         "kill",
         "restart",
         "tcm",
+        "backup",
     ),
 )